feat(query-stack) upgrade for new approach the application layer.
