+++ src/Services/Identity/Query/Infrastructure.EventBus/Consumers/UserPasswordChangedConsumer.cs
@@ -4,8 +4,8 @@ using Infrastructure.EventBus.Abstractions;
 
 namespace Infrastructure.EventBus.Consumers;
 
-public class UserPasswordChangedConsumer : Consumer<DomainEvent.PasswordChanged>
+public class UserPasswordChangedConsumer : Consumer<DomainEvent.UserPasswordChanged>
 {
-    public UserPasswordChangedConsumer(IInteractor<DomainEvent.PasswordChanged> interactor)
+    public UserPasswordChangedConsumer(IInteractor<DomainEvent.UserPasswordChanged> interactor)
         : base(interactor) { }
 }
